core: allowances: create abstract class AbstractAllowanceWithRanges and only leave core computation into MarecoAllowance and LinearAllowance
