+++ core/src/main/java/fr/sncf/osrd/envelope_sim/allowances/Allowance.java
@@ -1,11 +1,8 @@
 package fr.sncf.osrd.envelope_sim.allowances;
 
 import fr.sncf.osrd.envelope.Envelope;
+import fr.sncf.osrd.envelope_sim.allowances.utils.AllowanceValue;
 
 public interface Allowance {
     Envelope apply(Envelope base);
-
-    double getDistance();
-
-    double getAllowanceTime(double baseTime);
 }
